+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/button/BooleanNullableCheckBoxValue.java
@@ -24,7 +24,7 @@ final class BooleanNullableCheckBoxValue extends AbstractComponentValue<Boolean,
 
 	BooleanNullableCheckBoxValue(NullableCheckBox checkBox) {
 		super(checkBox);
-		checkBox.getNullableModel().addListener(value -> notifyListeners());
+		checkBox.getNullableModel().stateObserver().addListener(this::notifyListeners);
 	}
 
 	@Override